rename: goods_all_~ -> goods_~ 로 변경
- 서비스와 리포지토리의 경우 해당 도메인에서 공유하여 사용
